Fixed problem with zero NTP leading to 07:28
Issue: If the 32-bit time value in the NTP packet contains all zeroes, this is interpreted as 2036-02-07T06:28:18Z. Therefore, the clock is set to the wrong time (to 7:28 in CET).
In applications where a wake up call is used, if it is set to 7:30, the alarm will go off after two minutes, which is annoying.
Solution: check for the time value being 0. If so, ignore it.